+++ frontend/package.json
@@ -26,6 +26,7 @@
     "react-query": "^3.39.2",
     "react-router-dom": "^6.3.0",
     "react-scripts": "5.0.1",
+    "supergroup": "^1.1.9",
     "web-vitals": "^2.1.4"
   },
   "scripts": {
